Data Visualization Practice 3
For this assignment, display the database information in an interesting and visual way. More specifically, you have the data to create a Bar Graph Animation of the Positive Reviews for the restaurants, that sorts and filters the data for the user, while showing them the data in a visual way.